refactor : board manage 객체들을 별도 폴더에 분리
